+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ## [Unreleased]
 ### Fixed
 - Fixes a crash when opening the output selection (or any other API call) before a default connection is configured.
+- Fixes a crash when the plugin sends a cover protocol message with a null `cover` field.
 
 ## [1.6.0-rc.2] - 2026-04-18
 ### Fixed